sched: SCHED_HMP multi-domain task migration control
We need a way to prevent tasks that are migrating up and down the
hmp_domains from migrating straight on through before the load has
adapted to the new compute capacity of the CPU on the new hmp_domain.
This patch adds a next up/down migration delay that prevents the task
from doing another migration in the same direction until the delay
has expired.

Change-Id: If26e7dcfa03e8b82e083452527a52020de4e2e8d
Signed-off-by: Morten Rasmussen <Morten.Rasmussen@arm.com>