s5p-fimc: Align imagesize to row size for tiled formats
For tiled format, we need to allocate a multiple of the row size. A good
example is 1280x720, which gets adjusted to 1280x736. In tiles this means Y
plane is 20x23 and UV plane is 20x12. Because of the rounding, the previous
code would only have enough space to fit half of the last row.