drm/exynos: hdmi: fix dt broken issue
This patch resolves the issue that the use of existing dtb is broken.

This patch gets existing dt nodes directly and checks if existing dtb
should be supported or not.
For this, it adds a soc_ver varable to hdmi_context structure to distinguish
SoC version because it should get dt node corresponding to SoC correctly.

Signed-off-by: Inki Dae <inki.dae@samsung.com>