upstream: [media] dib7000p: Callibrate signal strength
Calibrate the signal strength to dB.

For this callibration, I used adapter 0 (antenna 1), connected
via a normal cable to a DTF-2111 generator.

The same cabling and RF generator connected to adapter 1
(antenna 2) has a difference of +3dBm (signal was stronger
there).

Yet, changing the RF for a difference of, let's say, 6dB
reflected on a 6dB difference at the measured signal, with
is good.

So, the signal strengh can indeed be used to measure the
antenna gain, if the antenna is repositioned. It is
not precise to measure the absolute dBm value, however.

Backports http://git.linuxtv.org/cgit.cgi/media_tree.git/commit/?id=3926d91a6b93
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>