au0828: fix checks if dvb is initialized
dev->dvb is always not null, as it is an area at the dev
memory. So, checking if (dev->dvb) is always true.

Instead of this stupid check, what the code wants to do is
to know if the DVB was successully registered.

Fix it by checking, instead, for dvb->frontend. It should
also be sure that this var will be NULL if the device was
not properly initialized.

Change-Id: I714b87027e254d34c0448ecb9a673ce1bd9e9e32
Signed-off-by: Mauro Carvalho Chehab <m.chehab@samsung.com>