HACK: drm/exynos: g2d: devfreq integration
Some debugging has revealed that the operating frequency of the
G2D engine has some restrictions with respect to the DMC bus
(or just the ACP subbus) frequency. The engine seems to be part
of the DMC bus domain and accesses memory through it.

With the current overclocking of the engine to 400Mhz, any
attempt to run DMC lower than also 400MHz results in:
- complete system lockup (kernel is no longer alive)
- IOMMU pagefault at random/weird addresses
- engine hang, which can be mitigated with a HW reset

This hack uses the devfreq subsystem to lock DMC to 400MHz
once the G2D device is opened, and restores it to its defaults
again on close.

This is still broken/unfinished:
- currently the devfreq device is requested in the open()
  call, and not while probing (Exynos/G2D is not able to
  probe defer in its current form)
- DMC has subdomains like ACP (we should check if it suffices
  to lock that subdomain to a minimum frequency)